pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "monarch-cli-sync"
version = "0.1.0"
description = "Sync Amazon order history into Monarch Money without a browser extension"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "amazon-orders",
    "monarchmoney>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "respx>=0.21",
    "responses>=0.25",
    "pytest-cov>=5.0",
]

[project.scripts]
monarch-cli-sync = "monarch_cli_sync.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["monarch_cli_sync"]
